Adding the weather data feature #108 (#112)
* Bugfix: The default format should use the selected unit

* Updating the commons-suncalc

* Bugfix: saving the folder when saving the MRB as an image #88 (#19)

* Modifying the %A and %D tags displays #

* Removing the M and km characters

* Feature: Adding the Thunderforest Outdoors layer #37 (#20)

* Removing the extra space #105 (#21)

* Adding the weather data feature #108

* Displaying the last 3 years of weather data

* Syncing the 3 bundles

* Adding the field description

* Resolving conflicts

* The API Key was not written in the config XML

* Assigning and displaying the temperature at each closest location

* Putting back what was ommited in the previous merge

* Placing the label and text fields correctly

* Creating a separate tab for the weather

* This will be for a "Part2" (see branch weather-part2)

* Adding the climacons images

* Displaying the right icons for each corresponding forecast data

* Adding the converted images. They may need to be cropped in the future.

* The correct icon is displayed accordingly

* Setting up the weather data report

* More progress for the weather report

* Adding a script to convert and crop the Climacons SVG

* Adding the day to the weather report

* Rearrangeing the weather report and displaying the thermometers

* WIP : Handling the serialization/deserialization of the historical
weather

* Modifying the script and adding the converted images needed

* WIP Serialization/Deserialization of the historical weathe data

* More progress on the serialization/deserialization. Getting close!

* Handling the historical weather setter and cleaning

* Fixing the historic weather deserialization

* Specifying the data types

* Specifying the data types (part2)

* When saving the weather report, we convert the image paths to their
base64 representaton

* Removing the precip acumulation

* Updating the html reports

* Fixing time zone issues

* Fixing a bug where the sunrise/sunset values were wrong

* Fixing the time conversion issue for a historical weather request

* Fixing temperature time bugs and rearrangeing the report

* We need the exact same weekday for the 3 previous years. Also, because
the daily weather time doesn't correspond to the course start time, we
don't print it in the report.

* Bugfix for when the API Key is not valid but the CGX has the weather
data

+++ src/course_generator/dialogs/frmTrackSettings.java
@@ -60,7 +60,7 @@ public class frmTrackSettings extends javax.swing.JDialog {
 	private static final long serialVersionUID = 4203790672206475952L;
 	private ResourceBundle bundle;
 	private boolean ok;
-	private Double timezone;
+	private int timezone;
 	private boolean summertime;
 	private CgSettings settings;
 	private JPanel jPanelButtons;
@@ -345,14 +345,14 @@ public class frmTrackSettings extends javax.swing.JDialog {
 				setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
 
 				if (res.valid) {
-					timezone = Double.valueOf(res.TimeZone);
+					timezone = res.TimeZone;
 					summertime = res.SummerTime;
 
 					Date date = Utils.DateTimetoSpinnerDate(res.Sunrise);
+
 					spinEndNightModel.setValue(date);
 					date = Utils.DateTimetoSpinnerDate(res.Sunset);
 					spinStartNightModel.setValue(date);
-
 				}
 			}
 		});
